+++ app/src/main/java/com/money/manager/ex/account/FilterDialogFragment.java
@@ -53,7 +53,7 @@ public class FilterDialogFragment
 
     private TransactionFilter mTransactionFilter;
     private Account mAccount;
-    private int mRecords;
+    private long mRecords;
 
     public FilterDialogFragment() {
         // Required empty public constructor
@@ -149,7 +149,7 @@ public class FilterDialogFragment
         // Default Account
 //        CheckBox defaultAccountCheckbox = (CheckBox) view.findViewById(R.id.defaultAccountCheckbox);
 //        AppSettings preferences = new AppSettings(getActivity());
-//        Integer defaultAccountId = preferences.getGeneralSettings().getDefaultAccountId();
+//        Long defaultAccountId = preferences.getGeneralSettings().getDefaultAccountId();
 //        defaultAccountCheckbox.setChecked(mAccount.getId().equals(defaultAccountId));
 
         // Number of records
